move openapi parsing into utils
