+++ src/hackingBuddyGPT/utils/openapi/openapi_converter.py
