+++ src/hackingBuddyGPT/utils/openapi/openapi_parser.py
