+++ src/hackingBuddyGPT/utils/openapi/yaml_assistant.py
@@ -36,7 +36,6 @@ class YamlFileAssistant:
             The current implementation is commented out and serves as a placeholder for
             integrating with OpenAI's API. Uncomment and modify the code as needed.
         """
-        """
         assistant = self.client.beta.assistants.create(
             name="Yaml File Analysis Assistant",
             instructions="You are an OpenAPI specification analyst. Use your knowledge to check "
@@ -87,4 +86,3 @@ class YamlFileAssistant:
 
         # The thread now has a vector store with that file in its tool resources.
         print(thread.tool_resources.file_search)
-        """
