+++ src/hackingBuddyGPT/utils/prompt_generation/information/pentesting_information.py
@@ -9,7 +9,7 @@ from typing import List
 
 import pandas
 
-from hackingBuddyGPT.usecases.web_api_testing.documentation.parsing import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
 from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
     PromptPurpose,
 )
